+++ src/common/js/configure.js
@@ -5,12 +5,15 @@
 import _ from 'lodash';
 
 import { ReactiveVar } from 'meteor/reactive-var';
+import { Tracker } from 'meteor/tracker';
 
 let _conf = {};
 
 Validity._conf = new ReactiveVar( _conf );
 
 Validity._defaults = {
+    effectEnd: 'effectEnd',
+    effectStart: 'effectStart',
     verbosity: Validity.C.Verbose.CONFIGURE
 };
 
@@ -35,3 +38,15 @@ Validity.configure = function( o ){
 
 _.merge( _conf, Validity._defaults );
 Validity._conf.set( _conf );
+
+// make sure effectStart and effectEnd field names are set
+Tracker.autorun(() => {
+    const effectStart = Validity.configure().effectStart;
+    if( !effectStart || !_.isString( effectStart )){
+        console.error( 'pwix:validity expects effectStart be a non empty string, got', effectStart );
+    }
+    const effectEnd = Validity.configure().effectEnd;
+    if( !effectEnd || !_.isString( effectEnd )){
+        console.error( 'pwix:validity expects effectEnd be a non empty string, got', effectEnd );
+    }
+});